Cleanups: bold post-heading colons, ffmpeg → ffplay/ffprobe as approp.
In several places, ffprobe or ffplay commands said ffmpeg in the command breakdown. This is because of me doing a sloppy cut & paste in a previous commit in this PR.

Previously, colons succeeding a bold 'Notes' mini-heading were not themselves also bold. Reto noted this, and the Chicago Manual of Style does indicate that such colons should also be in bold.