Command name change: 'View Format Info' to 'View ffmpeg subprogram info'
'View format info' seemed to me to be poorly named, what this command actually does is to give details of various libraries, muxers, encoders, etc. 'View ffmpeg subprogram info' may not itself be a totally correct nomenclature, but it's a bit closer to the practical meaning, I think.

Also: a couple of minor text changes.